pyproject.toml
[tool.poetry]
name = "opsipxeconfd"
version = "4.3.0.0"
description = "opsi pxe configuration service"
homepage = "https://www.opsi.org"
license = "AGPL-3.0"
maintainers = ["uib GmbH <info@uib.de>"]
authors = ["uib GmbH <info@uib.de>"]
include = [
    "run-opsipxeconfd",
    "opsipxeconfd_data/**/*"
]

[[tool.poetry.source]]
name = "uibpypi"
url = "http://pypi.uib.gmbh:8080/simple"


[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-opsi = "^4.3.0.3"
python-opsi-common = "^4.3.0.2"
mysqlclient = "^2.0"
distro = {platform = "linux", version = "^1.5.0"}
inotify = "^0.2.10"
configargparse = "^1.5.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2"
pytest-cov = "^4.0"
pyinstaller = "^5.5"
pylint = "^2.12"
perflint = "^0.7.0"
tan = "^22.8"
flake8 = "^4.0.1"
mypy = "^0.991"

[tool.poetry.scripts]
opsipxeconfd = "opsipxeconfd.__main__:main"

[tool.black]
line-length = 140
target-version = ['py39', 'py310']
use-tabs = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
